requirements.txt
flask
gunicorn

pytest
pytest-flask
pytest-cov

flask-sqlalchemy
Flask-WTF
Flask-Login
wtforms[email]

Flask-Migrate
flask_mail